Take advantage of the import system of javascript
The philosophy behind introducing mext is to extend class definitions
in-place which the language does not yet support. We, nevertheless,
wanted to keep using the import/export system of the language so this
change is quite essential. With the pattern introduce here, when we
import something, it means we either extend it or use it in the module
where it is imported.